pyproject.toml
[project]
name = "py-coding-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Paulescu", email = "plabartabajo@gmail.com" } 
]
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.42.0",
    "fire>=0.7.0",
    "gitingest>=0.1.2",
    "loguru>=0.7.3",
    "ollama>=0.4.5",
    "openai>=1.59.5",
    "python-dotenv>=1.0.1",
    "typer[all]>=0.15.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
